Change definitions of inline functions to extern
The inline semantics of C99 is badly broken, or at least the clang
interpretation thereof. What is completely natural to use in C++ is
a nightmare to use correclty in C99. Apparenlty, whoever designed
this had never heard of a COMDAT section and thought the successful
usage model prevalent in C++ could safely ignored.

In C++, your header contains:

   inline int foo() { ... }

And everything works fine. You can include this anywhere, the compiler
will drop a copy of the definition of foo in every object file,
typically putting the symbols in a special section that means "this
symbol is replicated" (COMDAT, derived from Fortran I believe) and the
linker will chose one. In any case, It Just Works (TM).

In C99, at least the way clang implements it, if you do this,
then you get unsatisfied symbols, because the definition of
foo() is dropped in no object file. Well, it depends, really. On
whether the compiler decides to inline. So it may build opt but not
debug for example.

In order to avoid that, you can mark your function as extern
inline. But in that case, the compiler drops a copy of the function in
every single object file, like in C++, except that it does not mark
them as special, so now you get duplicate symbols. Sometimes.

You can mark your function as static inline, but not only is it quite
verbose, it makes the code ugly too. You see, inline, static and
extern are all 6-character words, so they align nicely, something that
static inline breaks entirely. You may be tempted to force #define
inline to static inline, but then this breaks code that doesn't know
about that and writes static inline, because you can't say static
twice. Oh, and it also drops a new copy of your function in every
single translation unit, so the resulting binary is bloated with
identical copies of your function all over the place.

Believe it or not, the correct way to do things according to clang is:

- To define the function as 'inline' in the header
- To put a declaration of the function as 'extern' in the .c file.

You read that right. You have to do things backwards. The definition
is in the header file, the declaration has to be in the implementation
file to tell the stupid compiler where to drop the one and only copy
of your file that you want.

This is so entirely brain damaged that it brings brain damage to a
completely new level of ineptitude. To infinity and beyond. Twice.